+++ includes/system/class-l10n.php
@@ -362,21 +362,30 @@ class L10n {
 	public static function get_main_lang_code( $country ) {
 		if ( I18n::is_extension_loaded() && '[' !== substr( self::$countries[ $country ], 0, 1 ) ) {
 			$subtags = \ResourceBundle::create( 'likelySubtags', 'ICUDATA', false );
-			$country = \Locale::canonicalize( 'und_' . $country );
+			// First try
+			$locale  = $subtags->get( \Locale::canonicalize( 'und_' . $country ) );
+			if ( $locale ) {
+				return \Locale::getPrimaryLanguage( $locale );
+			}
+			// Second try
+			$locale  = $subtags->get( 'und_' . $country );
+			if ( $locale ) {
+				return \Locale::getPrimaryLanguage( $locale );
+			}
+			// Third try
 			$locale  = $subtags->get( $country );
 			if ( $locale ) {
 				return \Locale::getPrimaryLanguage( $locale );
-			} else {
-				\DecaLog\Engine::eventsLogger( IPLOCATOR_SLUG )->debug( $subtags->getErrorMessage(), [ 'code' => $subtags->getErrorCode() ] );
-				$locale = $subtags->get( 'und' );
-				if ( $locale ) {
-					return \Locale::getPrimaryLanguage( $locale );
-				} else {
-					\DecaLog\Engine::eventsLogger( IPLOCATOR_SLUG )->info( $subtags->getErrorMessage(), [ 'code' => $subtags->getErrorCode() ] );
-				}
 			}
+			// Fourth try
+			/*$locale  = $subtags->get( \Locale::canonicalize( $country ) );
+			if ( $locale ) {
+				return \Locale::getPrimaryLanguage( $locale );
+			}*/
+			// Fallback
+			return 'en_US';
 		}
-		return 'en';
+		return 'en_US';
 	}
 
 	/**